+++ apps/desktop/.eslintrc.cjs
@@ -1,4 +1,0 @@
-module.exports = {
-  extends: ["../../eslint.config.js", "next/core-web-vitals"],
-  ignorePatterns: ["dist"],
-};