+++ apps/mobile/app/select-date/layout.tsx
@@ -1,0 +1,14 @@
+import SelectDateHeader from "../../components/select-date/SelectDateHeader";
+
+export default function layout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <section>
+      <SelectDateHeader />
+      <div className="px-[2rem] py-[3rem]">{children}</div>
+    </section>
+  );
+}