+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,74 @@
 # desktop
 
+## 0.1.14
+
+### Patch Changes
+
+- @setaday/ui@1.2.3
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [e249990]
+  - @setaday/ui@1.2.2
+
+## 0.1.12
+
+### Patch Changes
+
+- @setaday/ui@1.2.1
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [e3e1e14]
+  - @setaday/ui@1.2.0
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [1244f25]
+  - @setaday/ui@1.1.7
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [998edb9]
+  - @setaday/ui@1.1.6
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [d11a536]
+  - @setaday/ui@1.1.5
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [cfe5ac1]
+  - @setaday/ui@1.1.4
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [0408ee8]
+  - @setaday/ui@1.1.3
+
+## 0.1.5
+
+### Patch Changes
+
+- 58482be: update dependencies
+- Updated dependencies [db7a77b]
+  - @setaday/ui@1.1.2
+
 ## 0.1.4
 
 ### Patch Changes