+++ packages/design-token/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @setaday/design-token
 
+## 2.1.1
+
+### Patch Changes
+
+- af5bf45: fix red token typo error
+
+## 2.1.0
+
+### Minor Changes
+
+- 4574bd0: add red color token
+
 ## 2.0.0
 
 ### Major Changes