+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @setaday/ui
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [af5bf45]
+  - @setaday/design-token@2.1.1
+
+## 1.2.2
+
+### Patch Changes
+
+- e249990: set mobile default button width to 100%
+
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [4574bd0]
+  - @setaday/design-token@2.1.0
+
+## 1.2.0
+
+### Minor Changes
+
+- e3e1e14: Add maxlength error handling logic in TextField
+
+## 1.1.7
+
+### Patch Changes
+
+- 1244f25: fix style import error
+
+## 1.1.6
+
+### Patch Changes
+
+- 998edb9: fix style injection error
+
+## 1.1.5
+
+### Patch Changes
+
+- d11a536: extract css bundle for ui components
+
+## 1.1.4
+
+### Patch Changes
+
+- cfe5ac1: change import/export structure
+
+## 1.1.3
+
+### Patch Changes
+
+- 0408ee8: change tsup bundling entrypoint
+
+## 1.1.2
+
+### Patch Changes
+
+- db7a77b: change export path
+
 ## 1.1.1
 
 ### Patch Changes